Matrix inversion cache
Matrix inversion is usually a costly computation and there may be some
benefit to caching
The inverse of a matrix rather than compute it repeatedly. Here we
create a special matrix which stores it's inverse. Inverse is computed
only once for the very first time. Subsequently the cached value is
return.